Add examples of JSON to send to content-store
These are based on the content-store's ContentItem model and the data
required by manuals-frontend for displaying manuals and sections,
as well as our understanding of the manuals content. We have tried
to make this format applicable to manuals other than HMRC's, so that
manuals-frontend has as little to do as possible to display both simpler
and more complex manuals.

We think that it will be acceptable for manuals-frontend to need to
fetch the manual from the content-store in order to display a section;
requiring all data for displaying a section page to be present on the
section itself would lead to unnecessary duplication and updates to
all sections when eg the manual's public_updated_at changes.

One exception we have made to this is to include the manual's title
in the section. This makes the section's manual property structure
consistent with the breadcrumbs items.