Validate JSON request examples against the schemas
These examples were converted by hand from the scraped JSON in
manuals-frontend, and as the schemas evolve they are very likely to
become invalid JSON and inconsistent with the schemas. Validating
the examples against the schemas ensures that this doesn't happen.

This commit also removes some invalid trailing commas and fixes the
manual slug in EIM25525.