WIP reuse the child_section_groups definition
json-schema does support this, but in a way which would have meant it reading
two files for every request.

json_schema also supports this, but is a bit more cleanly written (although it
has its own niggles), and lets you reuse a schema object such that it doesn't
need to re-read the file every time.